feat: refactor seed scripts
